pyproject.toml
[tool.poetry]
name = "project"
version = "0.0.1"
description = ""
authors = ["Anton Zotov <antzotik@gmail.com>"]
readme = "README.md"
packages = [{include = "CIFAR_CNN"}]

[tool.poetry.dependencies]

python = ">=3.10,<3.13"
numpy = "^1.26.1"
torch = {version = "^2.1.0+cu118", source = "pytorch"}
torchvision = {version = "^0.16.0+cu118", source = "pytorch"}
matplotlib = "^3.8.0"
wandb = "^0.15.12"
lightning = "^2.1.0"
hydra-core = "^1.3.2"


[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"